refactor: split friends screen into modular sections and relocate logic
